Deport TodoList object to utils module
