+++ self_wiki/utils.py
@@ -1,12 +1,11 @@
 """
 Some useful classes and functions related to self.wiki
 """
+import json
 import os
-from os.path import join as pjoin
+from os.path import join as pjoin, exists
 from datetime import datetime
 from typing import List, Optional
-
-from self_wiki import CONTENT_ROOT
 
 
 class RecentFileManager:
@@ -17,9 +16,9 @@ class RecentFileManager:
     DEFAULT_LIMIT = 20
 
     @classmethod
-    def get_recent_files(cls, directory: str = CONTENT_ROOT,
+    def get_recent_files(cls, directory: str,
                          limit=DEFAULT_LIMIT,
-                         wanted_extensions: List[str] = ('md',)) -> List[dict]:
+                         wanted_extensions: Optional[List[str]] = None) -> List[dict]:
         """
         Returns the list of files, sorted by modification time as a UNIX timestamp (recent first),
         with an optional :param limit:.
@@ -28,11 +27,13 @@ class RecentFileManager:
         :type directory: str
         :param limit: number of results to return
         :type limit: int
-        :param wanted_extensions: A list of file extensions we want.
+        :param wanted_extensions: A list of file extensions we want. If None, ['md'] is used.
         :type wanted_extensions: list
         :return: a dictionary list with, where each dict has the following keys: path, mtime
         """
         files = []
+        if not wanted_extensions:
+            wanted_extensions = ['md']
         for path, dirnames, filenames in os.walk(directory):
             dirnames[:] = [d for d in dirnames if d != '.git']  # remove git dir(s)
             for fname in filenames:
@@ -49,16 +50,16 @@ class RecentFileManager:
         sorted_files = sorted(files, key=lambda x: x['mtime'], reverse=True)
         return sorted_files[:limit]
 
-    def __init__(self, root: str = CONTENT_ROOT, wanted_extensions: List[str] = ('md',)):
-        self.__root = root
-        self.__r = RecentFileManager.get_recent_files(directory=root,
-                                                      limit=self.DEFAULT_LIMIT,
-                                                      wanted_extensions=wanted_extensions)
+    def __init__(self, root: str, wanted_extensions: Optional[List[str]] = None):
+        self._root = root
+        self._file_list = RecentFileManager.get_recent_files(directory=root,
+                                                             limit=self.DEFAULT_LIMIT,
+                                                             wanted_extensions=wanted_extensions)
 
     @property
     def root(self):
         "Returns the path we consider as root."
-        return self.__root
+        return self._root
 
     def re_scan(self, limit: Optional[int] = None, wanted_extensions: List[str] = ('md',)):
         """
@@ -68,9 +69,9 @@ class RecentFileManager:
         :param limit: limit the number of results to this
         :return:
         """
-        self.__r = RecentFileManager.get_recent_files(directory=self.root,
-                                                      limit=limit,
-                                                      wanted_extensions=wanted_extensions)
+        self._file_list = RecentFileManager.get_recent_files(directory=self.root,
+                                                             limit=limit,
+                                                             wanted_extensions=wanted_extensions)
 
     def update(self, path: str):
         """
@@ -79,11 +80,11 @@ class RecentFileManager:
 
         :param path: path to the file, relative to RecentFileManager.root, or not.
         """
-        if not path.startswith(self.__root):
-            path = pjoin(self.__root, path)
+        if not path.startswith(self._root):
+            path = pjoin(self._root, path)
         now = datetime.now()
         self.delete(path)
-        self.__r.insert(0, {'path': path, 'mtime': now.timestamp()})
+        self._file_list.insert(0, {'path': path, 'mtime': now.timestamp()})
 
     def get(self, limit: Optional[int] = None):
         """
@@ -96,9 +97,9 @@ class RecentFileManager:
                              " call list() yourself")
         if not limit:
             limit = self.DEFAULT_LIMIT
-        if len(self.__r) < limit:
-            limit = len(self.__r)
-        return list(self.__r[:limit])
+        if len(self._file_list) < limit:
+            limit = len(self._file_list)
+        return list(self._file_list[:limit])
 
     def delete(self, path: str):
         """
@@ -106,5 +107,62 @@ class RecentFileManager:
         :param path:
         :return:
         """
-        self.__r[:] = [d for d in self.__r
-                       if d.get('path') != path]
+        self._file_list[:] = [d for d in self._file_list
+                              if d.get('path') != path]
+
+class TodoList:
+    """
+    A container for a collection of Todos
+    """
+    def __init__(self, serialization_path):
+        "Creates a new TodoList collection"
+        self._todos = []
+        self._serialization_path = serialization_path
+        self.load()
+
+    def load(self):
+        """
+        load a serialized collection from disk
+        """
+        if not exists(self._serialization_path):
+            return
+        with open(self._serialization_path) as todo_file:
+            self._todos = json.load(todo_file)
+
+    def save(self):
+        """
+        persist current collection on disk
+        :return:
+        """
+        with open(self._serialization_path, 'w+') as f:
+            json.dump(self.todos, f)
+
+    def from_json(self, j: dict):
+        """
+        Insert an element from a dictionary object. Tries to compensate for eventual missing id.
+        :param j: A dictionary containing at least a 'text' key
+        """
+        if 'id' not in j.keys():
+            j['id'] = self._get_next_available_id()
+        else:
+            already_existing = list(filter(lambda x: x['id'] == j['id'], self._todos))
+            if already_existing is not None and already_existing != []:
+                already_existing[0].update(j)
+                return
+        self._todos.append(j)
+        self.save()
+
+    @property
+    def todos(self):
+        """
+        Returns the internal object list.
+
+        Why not rename self._todos to self.todos? No idea.
+        """
+        return self._todos
+
+    def _get_next_available_id(self):
+        current_ids_list = map(lambda x: x['id'], self._todos)
+        for i in range(0, 1024):  # totally arbitrary limit
+            if i not in current_ids_list:
+                return i
